// types.ts
export interface PersonalInfo {
  age: string;
  realName: string;
  occupation: string;
  hobbies: string;
}

export interface CustomImages {
  avatar: string;
  mainRobot: string;
  floatTopLeft: string;
  floatBottomLeft: string;
  floatTopRight: string;
  floatBottomRight: string;
}